+++ iothub_client/src/iothub_client_ll_uploadtoblob.c
@@ -745,7 +745,7 @@ IOTHUB_CLIENT_RESULT IoTHubClient_LL_UploadMultipleBlocksToBlob_Impl(IOTHUB_CLIE
                                         }
                                         else
                                         {
-                                            const size_t timeoutInMilliseconds = ((long)upload_data->blob_upload_timeout_secs * 1000);
+                                            const size_t timeoutInMilliseconds = (upload_data->blob_upload_timeout_secs * 1000);
                                             BLOB_RESULT uploadMultipleBlocksResult = Blob_UploadMultipleBlocksFromSasUri(STRING_c_str(sasUri), getDataCallbackEx, context, &httpResponse, responseToIoTHub, upload_data->certificates, &(upload_data->http_proxy_options), upload_data->networkInterface, timeoutInMilliseconds);
                                             if (uploadMultipleBlocksResult == BLOB_ABORTED)
                                             {
